pyproject.toml
[build-system]
requires = [
    "setuptools>=80.1.0",
    "setuptools_scm[toml]>=8.1",
    "wheel"
]
build-backend = 'setuptools.build_meta'

[project]
name = "STIXCore"
description = "An open-source Python data pipeline for the Spectrometer Telescope/Imaging X-rays (STIX) instrument on Solar Obriter."
requires-python = ">=3.11"
readme = {file = "README.rst", content-type="text/x-rst"}
license = "BSD-3-Clause"
license-files = ["LICENSE.rst"]
authors = [
    {name="Shane Maloney", email = "shane.maloney@dias.ie"},
    {name="Nicky Maloney", email = "shane.maloney@dias.ie"}
]
dependencies = [
    "bitstring~=4.2",
    "hissw~=2.3",
    "intervaltree",
    "numpy",
    "matplotlib",
    "paramiko",
    "polling2",
    "python-dateutil",
    "pyyaml",
    "roentgen~=2.1",
    "scp",
    "spiceypy~=6.0",
    "sunpy~=5.1",
    "watchdog~=6.0",
    "stixdcpy==3.0"
]
dynamic = ["version"]
keywords = ["solar physics", "solar", "sun", "solarorbiter", "stix"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Physics"
]

[project.optional-dependencies]
tests = [
    "pytest-astropy",
    "pytest-mpl",
    "pytest-xdist",
    "pytest>=7.1.0"
]
docs = [
    "sphinx",
    "sphinx-automodapi",
    "packaging>=23.0",
    "sunpy-sphinx-theme"
]
dev = ["stixcore[tests,docs]"]

[project.urls]
Homepage = "https://stixcore.readthedocs.io"
"Source Code" = "https://github.com/i4ds/STIXCore"
Documentation = "https://stixcore.readthedocs.io"
Changelog = "https://stixcore.readthedocs.io/en/latest/whatsnew/changelog.html"
"Issue Tracker" = "https://github.com/TCDSolar/i4ds/STIXCore"


[project.scripts]
stix-pipeline-status = "stixcore.processing.pipeline_status:main"
stix-pipeline-monitor = "stixcore.processing.pipeline_monitor:main"
stix-pipeline = "stixcore.processing.pipeline:main"
stix-pipeline-cli = "stixcore.processing.pipeline_cli:main"
stix-publish = "stixcore.processing.publish:main"

[tool.setuptools]
zip-safe = false
include-package-data = true
platforms = ["any"]
provides = ["stixcore"]

[tool.setuptools.packages.find]
include = ["stixcore*"]
exclude = ["stixcore._dev*"]

[tool.setuptools_scm.scm.git]
pre_parse = "fail_on_missing_submodules"  # Fail if submodules are not initialized

[tool.setuptools.package-data]
stixcore = ["data/*"]

[tool.setuptools_scm]
version_file = "stixcore/_version.py"

[tool.towncrier]
    package = "stixcore"
    filename = "CHANGELOG.rst"
    directory = "changelog/"
    issue_format = "`#{issue} <https://github.com/https://github.com/i4Ds/STIXCore.git/pull/{issue}>`__"

    [[tool.towncrier.type]]
        directory = "breaking"
        name = "Backwards Incompatible Changes"
        showcontent = true

    [[tool.towncrier.type]]
        directory = "api"
        name = "API Changes"
        showcontent = true

    [[tool.towncrier.type]]
      directory = "removal"
      name = "Deprecations and Removals"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "feature"
      name = "Features"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "bugfix"
      name = "Bug Fixes"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "doc"
      name = "Improved Documentation"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "trivial"
      name = "Trivial/Internal Changes"
      showcontent = true
